:art: refactor: useModal훅을 이용하여 useState 중복코드 줄이기
