+++ frontend/components/ui/modal/toggle-modal.interface.ts
@@ -3,7 +3,6 @@ import { IconType } from 'react-icons';
 export interface ToggleModalProps {
 	list: ToggleMenu[];
 	onClose: () => void;
-	modalWrapperRef: React.RefObject<HTMLDivElement>;
 }
 
 export interface ToggleMenu {
